requirements.txt
requests==2.31.0
pandas==1.3.5
pysftp==0.2.9
psycopg2-binary==2.9.6
